daraja/daraja/src/requirements.txt
asgiref==3.4.1
astroid==2.7.3
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
distlib==0.3.2
dj-database-url==0.5.0
Django==2.2.24
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.8.6
pylint==2.10.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.7
virtualenv==20.7.2
whitenoise==5.3.0
wrapt==1.12.1
